+++ packages/sil-governance/src/routes/create-proposal/index.tsx
@@ -6,6 +6,7 @@ import { useDispatch, useSelector } from "react-redux";
 
 import { history } from "..";
 import AsideFilter from "../../components/AsideFilter";
+import BlockchainTime from "../../components/BlockchainTime";
 import ConfirmTransaction from "../../components/ConfirmTransaction";
 import Header from "../../components/Header";
 import { RootState } from "../../store/reducers";
@@ -17,6 +18,7 @@ const CreateProposal = (): JSX.Element => {
   const dispatch = useDispatch();
   const lastSignAndPostResult = useSelector((state: RootState) => state.transactions.lastSignAndPostResult);
   const governor = useSelector((state: RootState) => state.extension.governor);
+  const blockchain = useSelector((state: RootState) => state.blockchain);
 
   const address = governor ? governor.address : ("" as Address);
   const [electorates, setElectorates] = useState<Readonly<Electorate[]>>([]);
@@ -45,7 +47,15 @@ const CreateProposal = (): JSX.Element => {
         <ConfirmTransaction transactionId={lastSignAndPostResult} onReturnToDashboard={onReturnToDashboard} />
       ) : (
         <Block minWidth="100%" display="flex">
-          <AsideFilter filter={null} />
+          <Block width="205px">
+            <AsideFilter filter={null} />
+            <Block marginTop={1} />
+            <Hairline />
+            <BlockchainTime
+              lastBlockHeight={blockchain.lastBlockheight}
+              lastBlockTime={blockchain.lastBlockTime}
+            />
+          </Block>
           <ProposalForm />
         </Block>
       )}