+++ packages/sil-governance/src/store/reducers.ts
@@ -1,12 +1,14 @@
 import { combineReducers, Reducer } from "redux";
 import { StateType } from "typesafe-actions";
 
+import { blockchainReducer, BlockchainState } from "./blockchain";
 import { extensionReducer, ExtensionState } from "./extension";
 import { proposalsReducer, ProposalsState } from "./proposals";
 import { transactionsReducer, TransactionsState } from "./transactions";
 
 export interface RootReducer {
   extension: ExtensionState;
+  blockchain: BlockchainState;
   proposals: ProposalsState;
   transactions: TransactionsState;
 }
@@ -14,6 +16,7 @@ export interface RootReducer {
 const createRootReducer = (): Reducer<RootReducer> =>
   combineReducers({
     extension: extensionReducer,
+    blockchain: blockchainReducer,
     proposals: proposalsReducer,
     transactions: transactionsReducer,
   });